requirements/base.txt
pytz==2022.7  # https://github.com/stub42/pytz
python-slugify==7.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.4.0  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django==4.1.5  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
djangorestframework-simplejwt==5.2.2 # https://django-rest-framework-simplejwt.readthedocs.io/
djangorestframework-api-key==2.2.0
opencv-python==4.7.0.68
numpy==1.24.1
django-cors-headers==3.13.0 # https://pypi.org/project/django-cors-headers/
drf-spectacular==0.25.1
